Remove Reaper Scans (Turkish)
