+++ multisrc/overrides/madara/reaperscans/src/ReaperScans.kt
@@ -1,0 +1,34 @@
+package eu.kanade.tachiyomi.extension.fr.reaperscans
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import eu.kanade.tachiyomi.source.model.SChapter
+import okhttp3.HttpUrl.Companion.toHttpUrl
+import org.jsoup.nodes.Element
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class ReaperScans : Madara(
+    "Reaper Scans",
+    "https://reaperscans.fr",
+    "fr",
+    SimpleDateFormat("dd/MM/yyyy", Locale.US),
+) {
+    // Migrated from WpMangaReader to Madara.
+    override val versionId = 2
+
+    override fun chapterFromElement(element: Element): SChapter = SChapter.create().apply {
+        val urlElement = element.selectFirst(chapterUrlSelector)!!
+
+        name = urlElement.selectFirst("p.chapter-manhwa-title")?.text()
+            ?: urlElement.ownText()
+        date_upload = urlElement.selectFirst("span.chapter-release-date > i")?.text()
+            .let { parseChapterDate(it) }
+
+        val fixedUrl = urlElement.attr("abs:href").toHttpUrl().newBuilder()
+            .removeAllQueryParameters("style")
+            .addQueryParameter("style", "list")
+            .toString()
+
+        setUrlWithoutDomain(fixedUrl)
+    }
+}
