+++ multisrc/src/main/java/eu/kanade/tachiyomi/multisrc/madara/MadaraGenerator.kt
@@ -15,7 +15,6 @@ class MadaraGenerator : ThemeSourceGenerator {
     override val sources = listOf(
         MultiLang("MangaForFree.net", "https://mangaforfree.net", listOf("en", "ko", "all"), isNsfw = true, className = "MangaForFreeFactory", pkgName = "mangaforfree", overrideVersionCode = 1),
         MultiLang("Manhwa18.cc", "https://manhwa18.cc", listOf("en", "ko", "all"), isNsfw = true, className = "Manhwa18CcFactory", pkgName = "manhwa18cc", overrideVersionCode = 4),
-        MultiLang("Reaper Scans", "https://reaperscans.com", listOf("fr", "tr"), className = "ReaperScansFactory", pkgName = "reaperscans", overrideVersionCode = 12),
         SingleLang("1st Kiss-Manga (unoriginal)", "https://1stkiss-manga.com", "en", className = "FirstKissDashManga"),
         SingleLang("1st Manhwa", "https://1stmanhwa.com", "en", isNsfw = true, className = "FirstManhwa"),
         SingleLang("1stKissManga.blog", "https://1stkissmanga.blog", "en", isNsfw = true, className = "FirstKissMangaBlog"),
@@ -198,6 +197,7 @@ class MadaraGenerator : ThemeSourceGenerator {
         SingleLang("Lolicon", "https://lolicon.mobi", "en", isNsfw = true, overrideVersionCode = 2),
         SingleLang("Lord Manga", "https://lordmanga.com", "en"),
         SingleLang("Luffy Manga", "https://luffymanga.com", "en"),
+        SingleLang("Lura Toon", "https://luratoon.com", "pt-BR", overrideVersionCode = 7, pkgName = "randomscan", className = "LuraToon"),
         SingleLang("LuxManga", "https://luxmanga.net", "en", isNsfw = true, overrideVersionCode = 1),
         SingleLang("MadaraDex", "https://madaradex.org", "en", isNsfw = true, overrideVersionCode = 1),
         SingleLang("Maid Scan", "https://maidscan.com.br", "pt-BR"),
@@ -415,13 +415,13 @@ class MadaraGenerator : ThemeSourceGenerator {
         SingleLang("RagnarokScan", "https://ragnarokscan.com", "es", overrideVersionCode = 1),
         SingleLang("Raijin Scans", "https://raijinscans.fr", "fr"),
         SingleLang("Rainbow Fairy Scan", "https://rainbowfairyscan.com", "pt-BR"),
-        SingleLang("Lura Toon", "https://luratoon.com", "pt-BR", overrideVersionCode = 7, pkgName = "randomscan", className = "LuraToon"),
         SingleLang("RawDEX", "https://rawdex.net", "ko", isNsfw = true, overrideVersionCode = 1),
         SingleLang("ReadAdult", "https://readadult.net", "en", isNsfw = true, overrideVersionCode = 1),
         SingleLang("ReaderGen", "https://fr.readergen.fr", "fr"),
         SingleLang("ReadFreeComics", "https://readfreecomics.com", "en"),
         SingleLang("ReadMangaFree", "https://readmangafree.net", "en", isNsfw = true),
         SingleLang("ReadManhua", "https://readmanhua.net", "en", overrideVersionCode = 2),
+        SingleLang("Reaper Scans", "https://reaperscans.fr", "fr"),
         SingleLang("Reset Scans", "https://reset-scans.us", "en"),
         SingleLang("Rh2PlusManga", "https://www.rh2plusmanga.com", "th", isNsfw = true, overrideVersionCode = 5),
         SingleLang("RichtoScan", "https://richtoscan.com", "es"),
